Create Summitive_Veer.py
My project idea is to use a website, to find the top 100 youtubers and calculate their average views per video. And then I am going to present that data using data visulisation, in a diagram. Today (26th September), I finished learning about the basics of python and web scrapping. I installed the files bs4, requesnts, and lxml, which are used for web scraping. My current code, gets my website to python and then prints all of the html text with the leter 'a' in it. Then it prints all of the text with 'name' in it. and then it prints the texts inside 'name'with 'href' in it. Currently, the code is slightly glitchy and it does not work. At home I will fix these errors