Create Summitive_Veer 2.py
Today (29th September) I tried to fix my code as it was not working last time. Now from the website I can load in the 'href' of the html's with 'a'. The data I need is in these strings. I am only confused on how pick out the even more specific data I want and how to convert the data into a list. Since I was unable to acomplish that, I worked on data visualisationa and with my code I can create an empty scatter plot graph with labels. Once I get my data lists than all I need to do is put the lists into the graph and I am done.