Create Summitive_Veer 3.py
Last time I was having a problem with selecting the specific data I want, however today I managed to remove all of the 'none' values I had in my data and also I removed some unesasry data. I also changed the names of my variables so that they make sense. The problem I am having now is that I am trying to split all of the strings in my list, but that is not working.