Create Summitive_Veer 4.py
previously, I managed to specify my data to only the values I wanted (the names of the youtube channels). However, the problem was that after each chanel name, there was 'page' writen and I couldn't figure out how to get rid of that as it was in the same string as all the names. Today I managed to fix that. I had to create an empty list, and then add then split all of the strings whenever there is a space. Then I added all of these values to the empty string.  But this was also problematic because now each chanel had a seprate list inside of the main list. To get rid of this, I created anouther empty string, and added only the channels to that string. In my project Idea, I had to first get the names of all of the youtubers from 1 website, and then I had to go to a 100 diffrent pages in that site to find the values I need. Now the problem is that loading all 100 pages at once is taking way to long.